Fixed minor type overflow on max zoom, showing time interval with mmbutton

+++ visualizer/src/TimeLineConfig.h
@@ -23,7 +23,7 @@ SOFTWARE. */
 #include <QColor>
 
 namespace visualizer {
-    using zoom_t = unsigned long int;
+    using zoom_t = unsigned long long int;
     using coord_t = long int;
 
     constexpr int       DefaultZoom             = 1000;
